fix(network-id): validate u32 truncation in service_send_raw_txn
`service.network_id` was stored as `u64` even though every consumer
(Solidity bridge `claimAssetCall.destinationNetwork`, `BridgeOutScanner`,
on-chain `networkID()` return) is `uint32`. The `params.destinationNetwork
!= service.network_id as u32` comparison in `handle_claim_asset`
truncated silently, so an operator who configured `--network-id`
> u32::MAX would have the proxy accept claims targeting
`network_id & 0xFFFFFFFF` — wrong network, silent acceptance.

Narrow once at startup (`main.rs`) via `u32::try_from(command.network_id)`,
mirroring the existing guard for `BridgeOutScanner`, and change
`ServiceState::network_id` to `u32` so no use site needs to cast. The
CLI flag stays `u64` for backward compat; an out-of-range value is now
a loud startup failure with a clear message rather than a silent
truncation at runtime.

Refs RD-703.

+++ src/service_send_raw_txn.rs
@@ -135,7 +135,14 @@ async fn handle_claim_asset(
     signer: Address,
 ) -> anyhow::Result<TxHash> {
     // Only claims where destinationNetwork matches our network_id are processed.
-    if params.destinationNetwork != service.network_id as u32 {
+    //
+    // RD-703 — `service.network_id` is `u32` (validated at startup in
+    // `main.rs` via `u32::try_from(command.network_id)`), matching the
+    // Solidity bridge's `uint32 destinationNetwork`. No silent `as u32` cast
+    // here: any operator value that does not fit `u32` is rejected loudly
+    // at startup rather than truncating into a comparison that would
+    // spuriously accept the wrong network.
+    if params.destinationNetwork != service.network_id {
         anyhow::bail!(
             "claim targets destinationNetwork {} but this proxy only handles network {}",
             params.destinationNetwork,
